fix(getQueryRunner): refactor and reduce tech debt
