+++ src/Components/ServiceScene/ServiceScene.tsx
@@ -20,7 +20,7 @@ import {
   SceneVariableValueChangedEvent,
   VariableDependencyConfig,
 } from '@grafana/scenes';
-import { VariableHide } from '@grafana/schema';
+import { LogsSortOrder, VariableHide } from '@grafana/schema';
 import { LoadingPlaceholder } from '@grafana/ui';
 
 import { plugin } from '../../module';
@@ -55,10 +55,13 @@ import { LEVELS_VARIABLE_SCENE_KEY, LevelsVariableScene } from '../IndexScene/Le
 import { ShowLogsButtonScene } from '../IndexScene/ShowLogsButtonScene';
 import { ActionBarScene } from './ActionBarScene';
 import { breakdownViewsDefinitions, valueBreakdownViews } from './BreakdownViews';
+import { getLogsPanelSortOrderFromURL } from './LogOptionsScene';
 import { LogsListScene } from './LogsListScene';
 import { drilldownLabelUrlKey, pageSlugUrlKey } from './ServiceSceneConstants';
+import { LokiQueryDirection } from 'services/lokiQuery';
 import { getQueryRunner, getResourceQueryRunner } from 'services/panel';
 import { buildDataQuery, buildResourceQuery } from 'services/query';
+import { getLogOption, getMaxLines } from 'services/store';
 import {
   DETECTED_FIELD_VALUES_EXPR,
   EMPTY_VARIABLE_VALUE,
@@ -805,11 +808,19 @@ function getDetectedFieldsQueryRunner() {
 }
 
 function getServiceSceneQueryRunner(sceneRef: SceneObject) {
-  return getQueryRunner(
-    [buildDataQuery(LOG_STREAM_SELECTOR_EXPR, { refId: LOGS_PANEL_QUERY_REFID })],
-    undefined,
-    sceneRef
-  );
+  const query = {
+    ...buildDataQuery(LOG_STREAM_SELECTOR_EXPR, { refId: LOGS_PANEL_QUERY_REFID }),
+    get direction() {
+      const sortOrder =
+        getLogsPanelSortOrderFromURL() || getLogOption<LogsSortOrder>('sortOrder', LogsSortOrder.Descending);
+      return sortOrder === LogsSortOrder.Descending ? LokiQueryDirection.Backward : LokiQueryDirection.Forward;
+    },
+    get maxLines() {
+      return getMaxLines(sceneRef);
+    },
+  };
+
+  return getQueryRunner([query], undefined);
 }
 
 function getLogCountQueryRunner() {
